Implemented class to convert notebooks to scripts
JupyterInvertor class takes the initial dataset directory and directory of refactored dataset. Firstly, it convert .ipynb file to json and then stores is as Notebook object. Notebook object has method to convert json representation of notebook to linear script. Thus, JupyterInvertor creates Notebook objects for all notebooks in projects and then creates corresponding scripts in the refactored dataset directory.

At the moment, the launch of this script via gradle is not implemented for the following reasons:

- JupyterInvertor class uses JSON, which downloads in build.gradle.kts, so it's impossible to put this class there.

- gradle doesn't see JupyterInvertor by import.

So I don't understand what to should I do to use it in Gradle.